backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0  
cryptography==44.0.0       
Django==5.1.3
django-cors-headers==4.6.0 
django-environ==0.11.2     
djangorestframework==3.15.2
idna==3.10
psycopg2-binary==2.9.10    
pycparser==2.22
PyJWT==2.10.1
pytz==2024.2
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
Pillow==11.0.0
